feat: restart stream if it crashes
